==> scripts/02_bamse_profiling.Rmd <==
---
title: "02_bamse_profiling"
output: html_document
autor: "María Bueno Álvez"
date: "2024-12-13"
editor_options: 
  chunk_output_type: console
---
 
# Set up

```{r}
source("scripts/functions/functions_analyses.R")
source("scripts/functions/functions_utility.R")
source("scripts/functions/functions_visualization.R")
source("scripts/functions/themes_palettes.R")

# Read in data
data_bamse <-
  read_tsv(
    "../Human-disease-blood-atlas/data/final_data/HPA/v24_2/bamse_data_ht_phase2.tsv"
  )
meta_bamse <-
  read_tsv(
    "../Human-disease-blood-atlas/data/final_data/HPA/v24_2/bamse_meta_ht_phase2.tsv"
  )
```


# Sampling overview

Visualize the sample collection timeline for each individual, with visit periods highlighted in the background.

```{r}
# Arrange subjects by first visit date
order_bamse <-
  meta_bamse |>
  filter(Visit == 4) |>
  arrange(Date)

# Define date ranges for each visit
visit_ranges <-
  meta_bamse %>%
  mutate(Visit = as.factor(Visit)) |>
  group_by(Visit) %>%
  summarize(start_date = min(Date), end_date = max(Date))

# Prepare plot of sample collection timeline per subject
meta_bamse |>
  mutate(Visit = as.factor(Visit),
         Subject = factor(Subject, levels = order_bamse$Subject)) |>
  ggplot(aes(Date, Subject, color = Visit, group = Subject)) +
  geom_rect(
    data = visit_ranges,
    aes(
      xmin = start_date,
      xmax = end_date,
      ymin = -Inf,
      ymax = Inf,
      fill = Visit
    ),
    alpha = 0.3,
    inherit.aes = FALSE
  ) +
  geom_line(show.legend = F, color = "grey90") +
  geom_point(show.legend = F) +
  scale_fill_manual(values = pal_bamse) +
  scale_color_manual(values = pal_bamse) +
  theme_hpa() +
  xlab("Sample date") +
  theme(axis.ticks.y = element_blank(),
        axis.text.y = element_blank())

ggsave(
  savepath_results("Manuscript-figures", "bamse_visit_timeline.pdf"),
  width = 7,
  height = 5
)
```

# Protein detectability analysiss

Analyze the number of proteins consistently detected above limit of detection (LOD).

## Calculate proteins detected above LOD in >80% of samples

```{r}
proteins_above_LOD_80 <-
  data_bamse |>
  mutate(above_LOD = ifelse(NPX > LOD, "Yes", "No")) |>
  group_by(Assay) |>
  count(above_LOD)  |>
  arrange(-n) |>
  mutate(n = (n / nrow(meta_bamse)) * 100) |>
  filter(above_LOD == "Yes", 
         n > 80)
```

## Remove proteins below LOD in > 80% of samples

```{r}
proteins_below_LOD_80 <-
  data_bamse |>
  mutate(above_LOD = ifelse(NPX > LOD, "Yes", "No")) |>
  group_by(Assay) |>
  count(above_LOD)  |>
  arrange(-n) |>
  mutate(n = (n / nrow(meta_bamse)) * 100) |>
  filter(above_LOD == "No",
         n > 80)

data_bamse_detected <-
  data_bamse |>
  filter(!Assay %in% proteins_below_LOD_80$Assay)
```

# UMAP

Visualize the stability and variation of the proteome across samples using dimensionality reduction.

## Generate UMAP

```{r}
umap_bamse <-
  do_umap(data = data_bamse_detected,
          wide = F,
          plots = F)
```

## Plot

```{r}
plot_umap <-
  umap_bamse |>
  left_join(meta_bamse, by = c("Sample" = "DAid")) |>
  mutate(Visit = factor(Visit)) |>
  ggplot(aes(UMAP1, UMAP2, color = Visit, fill = Visit)) +
  geom_point(alpha = 0.7) +
  stat_ellipse(geom = "polygon",
               alpha = 0.3,
               color = NA) +
  scale_color_manual(values = pal_bamse) +
  scale_fill_manual(values = pal_bamse) +
  theme_hpa() +
  ggtitle("UMAP")

ggsave(
  savepath_results("Publication-figures", "BAMSE_UMAP.pdf"),
  h = 5,
  w = 6
)
```
# Mixed-effects models

Estimate how much variation in protein levels is explained by individual differences (random effects) versus age and sex (fixed effects).

## Prepare data and run models 

```{r}
data_mm <-
  data_bamse_detected |>
  select(DAid, Assay, NPX) %>%
  left_join(meta_bamse, by = "DAid") |>
  mutate(Sex = as.factor(Sex),
         Age = as.factor(Age),
         Subject = as.factor(Subject))

# Run mixed effect models on all proteins extracting the fixed effects
fixed_effects <-
  map_df(unique(data_mm$Assay), function(protein) {
    do_mixed_effect_model(df = data_mm,
                          type = "fixed_effects",
                          protein = protein)
    
  })

# Run mixed effect models on all proteins extracting the variance explained
variance_explained <-
  map_df(unique(data_mm$Assay), function(protein) {
    do_mixed_effect_model(df = data_mm,
                          type = "variance_explained",
                          protein = protein)
  })
```
## Save results

```{r}
# Save results (RDS)
saveRDS(fixed_effects, 
        savepath_data("MM", "MM_estimates_BAMSE.rds"))
saveRDS(variance_explained,
        savepath_data("MM", "MM_variance_BAMSE.rds"))

# Save results
write_excel_csv(
  fixed_effects,
  savepath_data("Supplementary-files", "Data_S3_BAMSE_estimates.csv")
)
write_excel_csv(
  variance_explained,
  savepath_data("Supplementary-files", "Data_S3_BAMSE_variance.csv")
)
```

## Variance explained

```{r}
# Main
variance_explained |>
  filter(Component != "Residual") |>
  mutate(Variance = Variance * 100,
         Component = factor(
           Component,
           levels = c("Random effects (subject)", "Fixed effects (age & sex)")
         )) |>
  ggplot(aes(Variance, fill = Component)) +
  geom_histogram(show.legend = F) +
  facet_wrap( ~ Component, scales = "free") +
  scale_fill_manual(values = pal_mm) +
  xlab("Variance explained (%)") +
  ylab("Number of proteins") +
  theme_hpa()

ggsave(
  savepath_results("Manuscript-figures", "MM_variance_explained_BAMSE.pdf"),
  h = 3,
  w = 7
)

# Supplementary
order <-
  variance_explained |>
  filter(Component == "Residual") |>
  arrange(Variance)

variance_explained |>
  filter(Component != "Residual") |>
  mutate(Protein = factor(Protein, levels = order$Protein)) |>
  mutate(Variance = Variance * 100) |>
  ggplot(aes(Protein, Variance, fill = Component)) +
  geom_col(position = "stack") +
  scale_fill_manual(values = pal_mm) +
  theme_hpa() +
  theme(
    legend.position = "top",
    axis.ticks.x = element_blank(),
    axis.text.x = element_blank()
  ) +
  ylab("Variance explained (%)")

ggsave(
  savepath_results("Manuscript-figures", "Fig2_MM-variance-explained.pdf"),
  h = 5,
  w = 14
)
```

## Protein examples

```{r}
# Main
variance_explained |>
  filter(Component == "Random effects (subject)") |>
  arrange(-Variance)

data_bamse |>
  filter(Assay %in% c("CD33", "AMBN")) |>
  mutate(Assay = factor(Assay, levels = c("CD33", "AMBN"))) |>
  left_join(meta_bamse, by = "DAid") |>
  mutate(Visit = factor(Visit)) |>
  ggplot(aes(Visit, NPX, color = Sex, group = Subject)) +
  geom_point() +
  geom_line(show.legend = F, alpha = 7) +
  scale_color_manual(values = pal_sex) +
  facet_wrap( ~ Assay, nrow = 1, scale = "free") +
  theme_hpa() +
  theme(legend.position = "top")

ggsave(
  savepath_results("Manuscript-figures", "BAMSE-MM-example-individual.pdf"),
  h = 3.5,
  w = 6
)

# Supplementary
top_100 <-
  order |>
  mutate(Variance = Variance * 100) |>
  head(100)

p1 <-
  variance_explained |>
  filter(Protein %in% top_100$Protein) |>
  mutate(Protein = factor(Protein, levels = top_100$Protein)) |>
  filter(Component != "Residual") |>
  mutate(Variance = Variance * 100) |>
  ggplot(aes(Protein, Variance, fill = Component)) +
  geom_col(position = "stack") +
  scale_fill_manual(values = pal_mm) +
  theme_hpa(angled = T) +
  ylab("Variance explained (%)")

top_proteins <-
  variance_explained |>
  filter(Component == "Random effects (subject)") |>
  arrange(-Variance) |>
  head(5) |>
  bind_rows(
    variance_explained |>
      filter(Component == "Fixed effects (age & sex)") |>
      arrange(-Variance) |>
      head(5)
  ) |>
  mutate(Variance = Variance * 100)

proteins_random <-
  top_proteins |>
  filter(Component == "Random effects (subject)")

p2 <-
  data_bamse |>
  filter(Assay %in% proteins_random$Protein) |>
  mutate(Assay = factor(Assay, levels = proteins_random$Protein)) |>
  left_join(meta_bamse, by = "DAid") |>
  mutate(Subject = factor(Subject)) |>
  ggplot(aes(Visit, NPX, color = Subject)) +
  geom_point(show.legend = F) +
  geom_line(show.legend = F) +
  scale_color_manual(values = colorRampPalette(brewer.pal(12, "Paired"))(100)) +
  facet_wrap(~ Assay, nrow = 1, scale = "free") +
  theme_hpa()

p1 / p2
ggsave(
  savepath_results("Manuscript-figures", "MM-bamse-supl.pdf"),
  h = 6.6,
  w = 15
)
```

## Age

### Prepare

```{r}
dat_age <-
  fixed_effects |>
  filter(term == "Age") |>
  mutate(p.value = p.adjust(p.value, method = "BH")) |>
  arrange(p.value) |>
  mutate(
    sig = case_when(
      p.value < 0.06 & estimate > 0 ~ "significant up",
      p.value < 0.06 &
        estimate < 0 ~ "significant down",
      TRUE ~ "not significant"
    )
  )
```

### Volcano

```{r}
labels_age <-
  dat_age |>
  group_by(sig) |>
  top_n(n = 3, wt = -log10(p.value)) |>
  filter(sig != "not significant")

volcano_plot_age <-
  dat_age |>
  ggplot(aes(
    x = estimate,
    y = -log10(p.value),
    color = sig,
    label = Assay
  )) +
  geom_point(size = 1, alpha = 0.4) +
  geom_text_repel(data = labels_age,
                  size = 2,
                  show.legend = F) +
  geom_hline(
    yintercept = -log10(0.05),
    linetype = 'dashed',
    color = "darkgrey"
  ) +
  scale_color_manual(values = pal_de) +
  theme_hpa() +
  theme(axis.text = element_text(size = 8),
        legend.position = "top") 
```

### Proteins

```{r}
top_age_proteins <-
  dat_age |>
  group_by(sig) |>
  top_n(n = 3, wt = -log10(p.value)) |>
  filter(sig != "not significant")

boxplots_age <-
  data_bamse |>
  filter(Assay %in% top_age_proteins$Assay) |>
  left_join(meta_bamse |>
              select(DAid, Subject, Visit)) |>
  mutate(Assay = factor(Assay, levels = top_age_proteins$Assay),
         Visit = factor(Visit)) |>
  ggplot(aes(Visit, NPX, color = Visit)) +
  geom_line(aes(group = Subject),
            alpha = 0.3,
            color = "grey80") +
  geom_quasirandom(alpha = 0.7) +
  geom_boxplot(fill = NA,
               outlier.colour = NA,
               color = "grey20") +
  scale_color_manual(values = pal_bamse) +
  facet_wrap( ~ Assay, scales = "free_y", nrow = 2) +
  theme_hpa() +
  theme(
    legend.position = "top",
    axis.ticks.x = element_blank(),
    axis.text.x = element_blank()
  )
```

### Combine

```{r}
volcano_plot_age + boxplots_age

ggsave(
  savepath_results("Manuscript-figures", "Fig2c_Fig2d.pdf"),
  h = 5,
  w = 12
)
```

## Sex

### Prepare

```{r}
dat_sex <-
  fixed_effects |>
  filter(term == "SexM") |>
  mutate(p.value = p.adjust(p.value, method = "BH")) |>
  arrange(p.value) |>
  mutate(
    sig = case_when(
      p.value < 0.06 & estimate > 0 ~ "significant up",
      p.value < 0.06 &
        estimate < 0 ~ "significant down",
      TRUE ~ "not significant"
    )
  )
```

### Volcano

```{r}
labels_sex <-
  dat_sex |>
  group_by(sig) |>
  top_n(n = 3, wt = -log10(p.value)) |>
  filter(sig != "not significant")

volcano_plot_sex <-
  dat_sex |>
  ggplot(aes(
    x = estimate,
    y = -log10(p.value),
    color = sig,
    label = Assay
  )) +
  geom_point(size = 1, alpha = 0.4) +
  geom_text_repel(data = labels_sex,
                  size = 2,
                  show.legend = F) +
  geom_hline(
    yintercept = -log10(0.05),
    linetype = 'dashed',
    color = "darkgrey"
  ) +
  scale_color_manual(values = pal_de) +
  theme_hpa() +
  theme(axis.text = element_text(size = 8),
        legend.position = "top") 
```

### Proteins

```{r}
top_sex_proteins <- 
  dat_sex |> 
  group_by(sig) |>
  top_n(n = 3, wt = -log10(p.value)) |> 
  filter(sig != "not significant") 

boxplots_sex <- 
  data_bamse |> 
  filter(Assay %in% top_sex_proteins$Assay) |>
  left_join(meta_bamse |> 
              select(DAid, Subject, Visit, Sex)) |> 
  mutate(Assay = factor(Assay, levels = top_sex_proteins$Assay),
         Visit = factor(Visit)) |>
  ggplot(aes(Visit, NPX, color = Sex)) +
  geom_line(aes(group = Subject), alpha = 0.3, color = "grey80") +
  geom_quasirandom(alpha = 0.7) +
  scale_color_manual(values = pal_sex) +
  facet_wrap(~Assay, scales = "free_y", nrow = 2) +
  theme_hpa() +
  theme(legend.position = "top",
        axis.ticks.x = element_blank(),
        axis.text.x = element_blank())
```

### Combine

```{r}
volcano_plot_sex + boxplots_sex

ggsave(savepath_results("Manuscript-figures", "Fig2e_Fig2f.pdf"), h = 5, w = 12)
```
