==> scripts/04_validation.Rmd <==
---
title: "04_validation"
output: html_document
autor: "María Bueno Álvez"
date: "2024-12-13"
editor_options: 
  chunk_output_type: console
---

# Set up

```{r}
source("scripts/functions/functions_analyses.R")
source("scripts/functions/functions_utility.R")
source("scripts/functions/functions_visualization.R")
source("scripts/functions/themes_palettes.R")

# Read in data - UKB
ukb_data <- read_delim("data/UKBB_cancer_olink.txt")

ukb_meta <- read_delim("data/UKBB_cancer_meta.txt")
ukb_meta <-
  ukb_meta |> 
  distinct(eid, .keep_all = TRUE)

ukb_data <- 
  ukb_data |> 
  separate(meaning,  c("Assay", "Meaning")) |>
  select(eid, Assay, NPX = result)

# Read in data - UCAN
ucan_data <- read_tsv("data/data_phase2_batch2_curated_20250127.tsv") 
manifest <- import_df("data/samples_2024-12-17.xlsx")

# Read in data - pan-disease
hdba_data <- read_tsv("data/data_hdba_hpa_p1_v24.tsv")
hdba_meta <- read_tsv("data/meta_hdba_hpa_p1_v24.tsv")

olink_targets <-  import_df("data/overlap_olink_platforms.csv")
```

# UKB

## Load data exported from UKB-RAP

```{r}

```

## Time to diagnosis

```{r}
# Add n label
ukb_meta |> 
  mutate(time_to_diagnosis = Cancer_age_diagnosis - Age_sample_collect,
         Cancer_name = ifelse(Cancer_name %in% c("Colon_cancer", "Rectum_cancer"), "Colorectal_cancer", Cancer_name)) |> 
  ggplot(aes(time_to_diagnosis)) +
  geom_histogram(fill = pal_class["Cancer"]) +
  geom_vline(xintercept = 0) +
  scale_fill_viridis_c() +
  #geom_text(aes(label = n), vjust = -0.5) +
  facet_wrap(~Cancer_name, nrow = 1, scales = "free_y") +
  theme_hpa() +
  xlab("Time to diagnostis") +
  ylab("n")


ggsave(savepath("age_ukb.png"), h = 5, w = 15)


ukb_dat <- 
  ukb_meta |> 
  mutate(time_to_diagnosis = Age_sample_collect - Cancer_age_diagnosis,
         Cancer_name = ifelse(Cancer_name %in% c("Colon_cancer", "Rectum_cancer"), "Colorectal_cancer", Cancer_name),
         Group = case_when(time_to_diagnosis <= -7 ~ "> 7 years before",
                           time_to_diagnosis > -7 & time_to_diagnosis <= -5 ~ "5-7 years before",
                           time_to_diagnosis > -5 & time_to_diagnosis <= -3 ~ "3-5 years before",
                           time_to_diagnosis > -3 & time_to_diagnosis <= -1 ~ "1-3 years before",
                           time_to_diagnosis > -1 & time_to_diagnosis <= 1 ~ "1 year before/after",
                           time_to_diagnosis > 1 & time_to_diagnosis <= 3 ~ "1-3 years after",
                           time_to_diagnosis > 3 ~ "> 3 years after"),
         Group = factor(Group, levels = names(pal_ukb_2[-1]))) |> 
  filter(!time_to_diagnosis > 20) 

ukb_cancers <- c(
                 "Colorectal_cancer",
                 "Lung_cancer",
                 "Ovary_cancer",
                 #"Pancreas_cancer",
                 "Prostate_cancer"
                 )

ukb_dat |> 
  filter(Cancer_name %in% ukb_cancers) |> 
  mutate(Cancer_name = factor(Cancer_name, levels = ukb_cancers)) |> 
  ggplot(aes(time_to_diagnosis, fill = Group)) +
  geom_histogram() +
  geom_vline(xintercept = 0) +
  scale_fill_manual(values = pal_ukb_2) +
  facet_wrap(~Cancer_name, nrow = 1, scales = "free_y") +
  theme_hpa() +
  ggtitle("") +
  xlab("Time to diagnosis")
  

ggsave(savepath_results("Publication-figures","age_ukb.pdf"), h = 4.5, w = 16)
```
## Examples

```{r}
# Examples
selected_cancers <- 
  c("Lung cancer",
    "Colorectal cancer",
    "Pancreatic cancer",
    "Prostate cancer",
    "Ovarian cancer")

top_prots <- 
  combined_de |> 
  filter(Disease %in% selected_cancers,
         Control == "All other diseases") |> 
  group_by(Disease) |> 
  top_n(1, -adj.P.Val) |> 
  ungroup() |> 
  mutate(Cancer_name = case_when(Disease == "Pancreatic cancer" ~ "Pancreas_cancer",
                                  Disease == "Prostate cancer" ~ "Prostate_cancer",
                                  Disease == "Lung cancer" ~ "Lung_cancer",
                                  Disease == "Colorectal cancer" ~ "Colorectal_cancer"))  |> 
  filter(Assay != "MMP12")

resource_data %>%
  left_join(resource_meta |> 
              select(DAid, Disease, Class), by = "DAid") |> 
  filter(Assay %in% top_prots$Assay) %>% 
  mutate(Disease = factor(Disease, levels = disease_class_order)) |> 
  ggplot(aes(Disease, NPX, color = Class, fill = Class)) +
  geom_quasirandom( size = 0.5, show.legend = F, alpha = 0.7) +
  geom_boxplot(alpha = 0.3, color = "black", outlier.colour = NA, show.legend = F) +
  scale_color_manual(values = pal_class) +
  scale_fill_manual(values = pal_class) +
  facet_wrap(~Assay, nrow = 1, scales = "free_y") +
  xlab("") +
  ylab("Protein expression") +
  coord_flip() +
  theme_hpa(angled = T, axis_y = F) 
      
ggsave(savepath("examples_proteins.pdf"), h = 15, w = 10)


ukb_dat |> 
  left_join(ukb_meta, by = c("DAid" = "eid"))  |> 
  mutate(time_to_diagnosis = Age_sample_collect - Cancer_age_diagnosis,
         Cancer_name = ifelse(Cancer_name %in% c("Colon_cancer", "Rectum_cancer"), "Colorectal_cancer", Cancer_name),
         Group = case_when(time_to_diagnosis <= -7 ~ "> 7 years before",
                           time_to_diagnosis > -7 & time_to_diagnosis <= -5 ~ "5-7 years before",
                           time_to_diagnosis > -5 & time_to_diagnosis <= -3 ~ "3-5 years before",
                           time_to_diagnosis > -3 & time_to_diagnosis <= -1 ~ "1-3 years before",
                           time_to_diagnosis > -1 & time_to_diagnosis <= 1 ~ "1 year before/after",
                           time_to_diagnosis > 1 & time_to_diagnosis <= 3 ~ "1-3 years after",
                           time_to_diagnosis > 3 ~ "> 3 years after"),
         Group = factor(Group, levels = names(pal_ukb_2[-1]))) |> 
  inner_join(top_prots |> 
               select(Cancer_name, Assay),
             by = c("Cancer_name",
                    "Assay")
                      ) |> 
#  filter(Assay %in% top_prots$Assay) |> 
  ggplot(aes(Group, NPX, color = Group, fill = Group)) +
  geom_quasirandom(alpha = 0.8, size = 0.5) +
  geom_boxplot(outlier.color = NA, alpha = 0.4, color = "black", ) +
#  stat_summary(fun = mean, geom = "point", shape = 95, size = 6, color = "black") +
  facet_wrap(Cancer_name~Assay, scales = "free_y", nrow = 1) +
  scale_color_manual(values = pal_ukb_2) +
  scale_fill_manual(values = pal_ukb_2) +
  theme_hpa(angled = T, axis_x = F)+
  xlab("") 


ggsave(savepath("ukb_examples_proteins.pdf"), h = 3, w = 13)
```


## Differential expression

```{r}

a <- 
  ukb_data |>
  filter(!Assay == "HLA") |> 
  inner_join(ukb_dat |> 
  filter(Cancer_name == cancer) |> 
  select(eid, time_to_diagnosis)) |> 
    mutate(eid = as.character(eid))

somehow_dupps <- 
  a |> 
  count(eid) |> 
  arrange(-n) |> 
  filter(n > 2000)

cancer <- "Ovary_cancer"

de_ukb <- 
  map_df(c("Colorectal_cancer", "Lung_cancer", "Ovary_cancer", "Prostate_cancer"), function(cancer) {
  
  join_data <<- 
  ukb_data |>
  filter(!Assay == "HLA") |> 
  inner_join(ukb_dat |> 
  filter(Cancer_name == cancer) |> 
  select(eid, time_to_diagnosis)) |> 
    mutate(eid = as.character(eid)) |> 
 #mutate(Sex = factor(Sex)) |> 
  filter(time_to_diagnosis < 1,
         !eid %in% somehow_dupps$eid) |> 
  pivot_wider(names_from = "Assay", values_from = "NPX") |> 
    rename(DAid = eid)

do_limma_continuous(join_data = join_data,
                    variable = "time_to_diagnosis",
                    correct = c(),
                   correct_type = c(),
                    pval_lim = 0.05,
                    logfc_lim = 0) |> 
  mutate(Cancer_name = cancer) 
})

de_ukb |> count(Cancer_name)
```

## Volcano plots

```{r}

labels <- 
  de_ukb |> 
  group_by(Cancer_name) |> 
  top_n(6, -adj.P.Val) |> 
  filter(sig != " not significant")

de_ukb |> 
  ggplot(aes(x = logFC, y = -log10(adj.P.Val), color = sig, label = Assay)) +
  geom_point(size = 1, alpha = 0.4) + 
  geom_text_repel(data = labels, size = 2, show.legend = F) +
  geom_hline(yintercept = -log10(0.05), linetype = 'dashed', color = "darkgrey") +
 # geom_vline(xintercept = -cutoff, linetype = 'dashed', color = "darkgrey") +
#  geom_vline(xintercept = cutoff, linetype = 'dashed', color = "darkgrey") +
  scale_color_manual(values = pal_de) +
  facet_wrap(~Cancer_name, scales = "free", nrow = 1) +
  theme_hpa() +   
  theme(axis.text = element_text(size = 8),
        legend.position = "right") +
  xlab("Model coefficient")

ggsave(savepath_results("Publication-figures", "ukb_de.pdf"), h = 3, w = 12)
```

## Proteins

```{r}

plot_nefl <- boxplot_ukb("Lung_cancer", "NEFL") 
#plot_cxcl8 <- boxplot_ukb("Colorectal_cancer", "CXCL8") 
plot_ceacam5 <- boxplot_ukb("Lung_cancer", "CEACAM5") 
plot_areg <-   boxplot_ukb("Colorectal_cancer", "AREG")
#plot_areg_lung <-   boxplot_ukb("Lung_cancer", "AREG")
plot_crc <- boxplot_ukb("Colorectal_cancer", "C1QA")
#plot_areg_lung <-   boxplot_ukb("Lung_cancer", "AREG")
#boxplot_ukb("Prostate_cancer", "KLK4")
plot_ovary <- boxplot_ukb("Ovary_cancer", "WFDC2")

boxplot_ukb("Prostate_cancer", "AREG")

plot_ceacam5 | plot_nefl | plot_crc | plot_areg | plot_ovary
ggsave(savepath_results("Publication-figures", "examples_ukp.pdf"), h = 3, w = 14)

protein <- 
cancer <- 

protein <- "CXCL8"
cancer <- "Colorectal_cancer"

protein <- "VTCN1"
cancer <- "Ovary_cancer"



ggsave(savepath("TFRC.png"), h = 4, w = 5)
ggsave(savepath("MDK_mockup.png"), h = 3, w = 4)
```


## Plot two examples

```{r}
protein <- "AGR2"

hdba_data |> 
  left_join(hdba_meta, by = "DAid") |>
  filter(Assay == protein,
         Disease %in% c("Ovarian cancer", 
                        "Breast cancer", 
                        "Prostate cancer",
                        "Lung cancer",
                        "Colorectal cancer")) |> 
  ggplot(aes(Disease, NPX, color = Disease, fill = Disease)) +
  geom_quasirandom() +
  geom_boxplot(outlier.color = NA, alpha = 0.4, color = "black") +
  scale_color_manual(values = (pal_ucan_2)) +
  scale_fill_manual(values = (pal_ucan_2)) +
  theme_hpa(axis_x = F)

ggsave(savepath("AGR2_P1.png"), h = 4, w = 5)
ggsave(savepath("CEACAM5_P1.png"), h = 4, w = 5)

ucan_data |> 
  left_join(manifest, by = "DAid") |>
  filter(Assay == protein,
         Diagnose %in% names(pal_ucan)) |> 
  ggplot(aes(Diagnose, NPX, color = Diagnose, fill = Diagnose)) +
  geom_quasirandom() +
  geom_boxplot(outlier.color = NA, alpha = 0.4, color = "black") +
  scale_color_manual(values = (pal_ucan)) +
  scale_fill_manual(values = (pal_ucan)) +
  theme_hpa(axis_x = F)

ggsave(savepath("CEACAM5_P2.png"), h = 4, w = 5)
ggsave(savepath("AGR2_P2.png"), h = 4, w = 5)





protein <- "PAEP"

hdba_data |> 
  left_join(hdba_meta, by = "DAid") |>
  filter(Assay == protein,
         Disease %in% c("Ovarian cancer", 
                        "Breast cancer", 
                        "Prostate cancer",
                        "Lung cancer",
                        "Colorectal cancer")) |> 
  ggplot(aes(Disease, NPX, color = Disease, fill = Disease)) +
  geom_quasirandom(show.legend = F) +
  geom_boxplot(outlier.color = NA, alpha = 0.4, color = "black", show.legend = F) +
  scale_color_manual(values = (pal_ucan_2)) +
  scale_fill_manual(values = (pal_ucan_2)) +
  theme_hpa() +
  theme(axis.ticks = element_blank(),
        axis.text = element_blank()) +
  xlab("Cancers") +
  ylab("Protein expression")

ggsave(savepath("PAEP_P1.png"), h = 4, w = 4)

ucan_data |> 
  left_join(manifest, by = "DAid") |>
  filter(Assay == protein,
         Diagnose %in% names(pal_ucan)) |> 
  ggplot(aes(Diagnose, NPX, color = Diagnose, fill = Diagnose)) +
  geom_quasirandom(show.legend = F) +
  geom_boxplot(outlier.color = NA, alpha = 0.4, color = "black", show.legend = F) +
  scale_color_manual(values = (pal_ucan)) +
  scale_fill_manual(values = (pal_ucan)) +
  theme_hpa() +
  theme(axis.ticks = element_blank(),
        axis.text = element_blank()) +
  xlab("Cancers") +
  ylab("Protein expression")

ggsave(savepath("PAEP_P2.png"), h = 4, w = 4)

```

# UCAN


## Number of samples


```{r}
cancers <- c("Breast cancer",
             "Colorectal cancer", 
             "Lung cancer", 
             "Ovarian cancer",
             "Prostate cancer")
ucan_meta <- 
  hdba_meta |> 
  filter(Disease %in% cancers)

p1 <- 
  ucan_meta |> 
  count(Disease) |> 
  mutate(Disease = case_when(Disease == "Breast cancer" ~ "BRC", 
                             Disease == "Colorectal cancer" ~ "CRC", 
                             Disease == "Lung cancer" ~ "LUNG", 
                             Disease == "Ovarian cancer" ~ "OVC", 
                             Disease == "Prostate cancer" ~ "PRC")) |> 
  mutate(Disease = factor(Disease, levels = c("LUNG", "CRC", "PRC", "BRC", "OVC"))) |> 
  ggplot(aes(Disease, n, fill = Disease)) +
  geom_col(show.legend = F) +
  geom_text(aes(label = n), vjust = 0.5) +
  coord_flip() +
  scale_y_reverse(limits = c(500, 0)) +  
  scale_fill_manual(values = pal_ucan) +
  xlab("") +
  theme_hpa(axis_y = F)

 manifest |> 
  filter(DAid %in% ucan_data$DAid) |> 
  filter(Diagnose != "bridge_sample") |> 
  count(Diagnose) |> pull(n) |> sum() 
 
p2 <- 
  manifest |> 
  filter(DAid %in% ucan_data$DAid) |> 
  filter(Diagnose != "bridge_sample") |> 
  count(Diagnose) |> 
  mutate(Diagnose = factor(Diagnose, levels = c("LUNG", "CRC", "PRC", "BRC", "OVC"))) |> 
  ggplot(aes(Diagnose, n, fill = Diagnose)) +
  geom_col(show.legend = F) +
  coord_flip() +
  scale_y_continuous(limits = c(0, 500)) +  # Ensure same scale
  geom_text(aes(label = n), vjust = 0.5) +
  theme_hpa() +
  scale_fill_manual(values = pal_ucan) +
  xlab("")

p1 + p2
ggsave(savepath("samples_ucan.pdf"), h = 3, w = 9)
```

## Differential expression


```{r}
female_diseases <- c("OVC", "BRC")
male_diseases <- c("PRC")

de_ucan <- 
  map_df(names(pal_ucan), function(cancer) {
    
    message("Processing disease: ", cancer)

    # Combine disease and control metadata
    disease_meta <- 
      manifest |> 
      filter(Cohort == "UCA2",
             Diagnose %in% names(pal_ucan)) |> 
      mutate(Disease = ifelse(Diagnose == cancer, cancer, "Control")) |> 
      select(DAid = DAid, Disease, Sex, Age, BMI)
    
    
    # Filter controls based on gender-specific diseases
    if(cancer %in% female_diseases) {
      
      disease_meta <- 
        disease_meta |> 
        filter(Sex == "F")
      
    } else if (cancer %in% male_diseases) {
      
      disease_meta <- 
        disease_meta |> 
        filter(Sex == "M")
      
    } else {
      
      disease_meta <- 
        disease_meta
      
    }
    
    disease_data <- 
      ucan_data |> 
      filter(DAid %in% disease_meta$DAid) |> 
      select(DAid, Assay, NPX) |> 
      pivot_wider(names_from = "Assay", values_from = "NPX")
    
    disease_res <- 
      do_limma_disease(data_wide = disease_data, 
                       metadata = disease_meta,
                       disease = cancer,
                       controls = "Cancers",
                       cutoff = 0.5,
                       correct = F)
    
  })
```

### Summary

```{r}
de_ucan_complete <- 
  map_df(unique(de_ucan$Disease), function(cancer) {
  
    sig_ucan_p1 <- 
  de_ucan_p1 |> 
  mutate(Cancer = ifelse(Cancer == "LUNGC", "LUNG", Cancer)) |> 
  filter(Cancer == cancer) |> 
  filter(Significance != "not significant")
  
  de_ucan |> 
  filter(Disease == cancer) |>
  left_join(olink_targets, by = "Assay") |> 
     mutate(Type = case_when(Assay %in% sig_ucan_p1$Assay & sig != "not significant" ~ "Replicated from Explore 1.5K", 
                          Platform != "1.5K" & sig != "not significant" ~ "New from Explore HT",
                          T ~ "Other")) 
  
})

de_ucan_complete |> 
  group_by(Disease) |> 
  count(Type) |> 
  filter(Type != "Other") |> 
  ggplot(aes(Type, n, fill = Type)) +
  geom_col() +
  geom_text(aes(label = n), vjust = 0.5) +
  coord_flip() +
  scale_color_manual(values = c("#E6998D", "#5F4A8B")) +
  scale_fill_manual(values = c("#E6998D", "#5F4A8B")) +
  facet_wrap(~Disease, nrow = 1) +
  theme_hpa()

ggsave(savepath("summary_ucan.pdf"), h = 2, w = 14)

```

### Top 1

```{r}
top1_targets <- 
  de_ucan_complete |>
  filter(sig == "significant up") |> 
  filter(Type == "New from Explore HT") |> 
  filter(Assay != "COL26A1") |> 
  group_by(Disease) |> 
  top_n(1, -adj.P.Val) |> 
  arrange(Disease) |> 
  mutate(Disease = factor(Disease)) |> 
  select(Assay, Disease)


ucan_data |>
  filter(NPX > LOD) |> 
  left_join(manifest |> 
              select(DAid, Diagnose), by = "DAid") |> 
  filter(Diagnose %in% unique(de_ucan_complete$Disease)) |> 
  inner_join(top1_targets) |> 
  mutate(Assay = factor(Assay, levels = top1_targets$Assay),
         Cancer = ifelse(Diagnose == Disease, Diagnose, "Other")) |> 
  ggplot(aes(Diagnose, NPX, fill = Cancer, color = Cancer)) +
  geom_quasirandom(alpha = 0.7, show.legend = F) +
  geom_boxplot(
    alpha = 0.3,
    outlier.color = NA,
    color = "grey20",
    show.legend = F
  ) +
  scale_color_manual(values = c(pal_ucan, "Other" = "grey")) +
  scale_fill_manual(values = c(pal_ucan, "Other" = "grey")) +
  facet_wrap(~Assay, scales = "free_y", nrow = 1) +
  theme_hpa(angled = T) +
  xlab("") 

ggsave(savepath("top1_ucan2.pdf"), h = 3, w = 11)

```



## Volcano plots - olink platform

```{r}
de_ucan_p1 <- read_tsv("data/ucan_p1.tsv")

volcanos_cancers <- 
  map(c("BRC", "CRC", "LUNG", "OVC", "PRC"), function(cancer) {
  sig_ucan_p1 <- 
  de_ucan_p1 |> 
  mutate(Cancer = ifelse(Cancer == "LUNGC", "LUNG", Cancer)) |> 
  filter(Cancer == cancer) |> 
  filter(Significance != "not significant")
  

labels <- 
  de_ucan |> 
  filter(Disease == cancer) |>
  left_join(olink_targets, by = "Assay") |> 
     mutate(Type = case_when(Assay %in% sig_ucan_p1$Assay & sig != "not significant" ~ "Replicated from Explore 1.5K", 
                          Platform != "1.5K" & sig != "not significant" ~ "New from Explore HT",
                          T ~ "Other")) |>
  group_by(Disease) |>
  top_n(n = 10, wt = -log10(adj.P.Val))

de_ucan |> 
  filter(Disease == cancer) |>
  left_join(olink_targets, by = "Assay") |> 
  mutate(Type = case_when(Assay %in% sig_ucan_p1$Assay & sig != "not significant" ~ "Replicated from Explore 1.5K", 
                          Platform != "1.5K" & sig != "not significant" ~ "New from Explore HT",
                          T ~ "Other")) |>
  mutate(Type = factor(Type, levels = c("Replicated from Explore 1.5K", "New from Explore HT", "Other"))) |> 
  ggplot(aes(x = logFC, y = -log10(adj.P.Val), color = Type, label = Assay)) +
  geom_point(size = 1, alpha = 0.4) + 
  geom_text_repel(data = labels, size = 2, show.legend = F) +
  geom_hline(yintercept = -log10(0.05), linetype = 'dashed', color = "darkgrey") +
  geom_vline(xintercept = -cutoff, linetype = 'dashed', color = "darkgrey") +
  geom_vline(xintercept = cutoff, linetype = 'dashed', color = "darkgrey") +
  scale_color_manual(values = c("#E6998D", "#5F4A8B", "grey")) +
  facet_wrap(~Disease, scales = "free", nrow = 1) +
  theme_hpa() +   
  theme(axis.text = element_text(size = 8),
        legend.position = "top") 
})

volcanos_cancers[[1]] | volcanos_cancers[[2]] | volcanos_cancers[[3]] | volcanos_cancers[[4]] | volcanos_cancers[[5]]

ggsave(savepath("ucan_volcanos.pdf"), h = 4.5, w = 15)
```

## Comparison to P1

```{r}
de_ucan_p1 <- read_tsv("data/previous_studies/ucan_p1.tsv")

de_p1_p2 <- 
  de_ucan_p1 |> 
  filter(Cancer %in% c("BRC", "CRC", "LUNGC", "OVC", "PRC")) |> 
  mutate(Cancer = ifelse(Cancer == "LUNGC", "LUNG", Cancer)) |> 
  select(Assay, Cancer, logFC_P1 = NPX_difference, adj_pval_P1 = p.adjusted) |>
  left_join(de_ucan |> 
               select(Assay, Cancer = Disease, logFC_P2 = logFC, adj_pval_P2 = adj.P.Val), by = c("Assay", "Cancer")) 

p_fc <- de_p1_p2 |> 
  ggplot(aes(x = logFC_P1, y = logFC_P2, color = Cancer)) +
  geom_point() +
  facet_wrap(~Cancer, scales = "free", nrow = 1) +
  geom_label_repel(aes(label = Assay), show.legend = F) +
  geom_hline(yintercept = 0, color = "grey", lty = "dashed") +
  geom_vline(xintercept = 0, color = "grey", lty = "dashed") +
  scale_color_manual(values = pal_ucan) +
  theme_hpa() 

p_pval <- 
  de_p1_p2 |> 
  ggplot(aes(x = -log10(adj_pval_P1), y = -log10(adj_pval_P2), color = Cancer)) +
  geom_point() +
  facet_wrap(~Cancer, scales = "free", nrow = 1) +
  geom_label_repel(aes(label = Assay), show.legend = F) +
  scale_color_manual(values = pal_ucan) +
  geom_hline(yintercept = -log10(0.05), color = "grey", lty = "dashed") +
  geom_vline(xintercept = -log10(0.05), color = "grey", lty = "dashed") +
  theme_hpa()

p_fc / p_pval

ggsave(savepath("p1_p2_ucan.png"), h = 10, w = 16)
```

## Examples

```{r}
proteins <- c("EGFR",
              "LGALS4",
              "CXCL17", 
              "PAEP",
              "FAP")

# HT
plot_ht <- 
  filtered_ucan_final |> 
    left_join(manifest, by = "DAid") |> 
    filter(Diagnose %in% names(pal_ucan),
           Assay %in% proteins) |> 
    mutate(Assay = factor(Assay, levels = proteins)) |> 
  filter(NPX > LOD) |> 
    ggplot(aes(Diagnose, NPX, fill = Diagnose, color = Diagnose)) +
    geom_quasirandom(alpha = 0.7) +
    geom_boxplot(
      alpha = 0.3,
      outlier.color = NA,
      color = "grey20"
    ) +
    scale_color_manual(values = pal_ucan) +
    scale_fill_manual(values = pal_ucan) +
    facet_wrap(~Assay, scales = "free_y", nrow = 1) +
    theme_hpa(angled = T, axis_x = F) +
    xlab("") +
  ggtitle("Olink HT")


# 1.5
data_NPX <- read_csv(file = "data/processed/final_data/data_phase1_20240604.csv")  

pal_ucan_p1 <- c("LUNGC" = "#ADC74F",
                 "CRC" = "#B89B74", 
                 "BRC" = "#E8A29A", 
                 "OVC" = "#603479", 
                 "PRC" = "#E7662B" )

data_ucan_p1 <- 
  data_NPX |> 
  left_join(manifest, by = "DAid") |> 
  filter(Cohort == "UCAN",
         Diagnose %in% names(pal_ucan_p1))


plot_p1 <- 
  data_ucan_p1 |> 
    filter(Assay %in% proteins,
           NPX > LOD) |> 
    mutate(Assay = factor(Assay, levels = proteins)) |> 
    ggplot(aes(Diagnose, NPX, fill = Diagnose, color = Diagnose)) +
    geom_quasirandom(alpha = 0.7) +
    geom_boxplot(
      alpha = 0.3,
      outlier.color = NA,
      color = "grey20"
    ) +
    scale_color_manual(values = pal_ucan_p1) +
    scale_fill_manual(values = pal_ucan_p1) +
    facet_wrap(~Assay, scales = "free_y", nrow = 1) +
    theme_hpa(angled = T, axis_x = F) +
    xlab("") +
    ggtitle("Olink Explore 1.5K")

plot_p1 /plot_ht 

ggsave(savepath("examples.png"), h = 7, w = 10)
```





## Machine learning

### Phase 1

```{r}
# Lists of diseases
female_diseases <- c("Ovarian cancer", "Breast cancer")
male_diseases <- c("Prostate cancer")

all_cancers <- 
  ucan_meta |> 
  distinct(Disease) |> 
  pull()

# Prepare data
ucan_p1_wide <- 
  hdba_data |> 
  filter(DAid %in% ucan_meta$DAid) |> 
  select(DAid, Assay, NPX) |> 
  pivot_wider(names_from = Assay, values_from = NPX) |>
  left_join(ucan_meta |> 
              select(DAid, Disease), by = "DAid")

# Split data
ucan_p1_split <- 
  generate_split(data = ucan_p1_wide, 
               proportion = 0.7,
               seed = 213,
               variable_stratify = "Disease")

# Run for all cancers
ucan_p1_ml <- 
  map(all_cancers, ~ do_lasso_binary(disease = .x,
                                     metadata = ucan_meta,
                                     split_train = ucan_p1_split$data_train, 
                                     split_test = ucan_p1_split$data_test))

saveRDS(ucan_p1_ml, savepath_data("ML", "ucan_p1_ml.rds"))


ovc_p1 <- do_lasso_binary(disease = "Ovarian cancer",
                          metadata = ucan_meta,
                          split_train = ucan_p1_split$data_train, 
                          split_test = ucan_p1_split$data_test) 

ovc_p1$important_proteins 

```

### Phase 2

```{r}
# Prepare metadata Olink HT
ucan_meta_p2 <- 
  manifest |> 
  filter(DAid %in% ucan_data$DAid) |> 
  filter(Diagnose != "bridge_sample") |>
  mutate(Disease = case_when(Diagnose == "BRC" ~ "Breast cancer", 
                             Diagnose == "CRC" ~ "Colorectal cancer", 
                             Diagnose == "LUNG" ~ "Lung cancer", 
                             Diagnose == "OVC" ~ "Ovarian cancer", 
                             Diagnose == "PRC"~ "Prostate cancer"))  
# Prepare data
ucan_p2_wide <- 
  ucan_data |> 
  filter(DAid %in% ucan_meta_p2$DAid) |> 
  select(DAid, Assay, NPX) |> 
  pivot_wider(names_from = Assay, values_from = NPX) |>
  left_join(ucan_meta_p2 |> 
              select(DAid, Disease), by = "DAid")

# Split data
ucan_p2_split <- 
  generate_split(data = ucan_p2_wide, 
               proportion = 0.7,
               seed = 213,
               variable_stratify = "Disease")

# Run for all cancers
ucan_p2_ml <- 
  map(all_cancers, ~ do_lasso_binary(disease = .x,
                                     metadata = ucan_meta_p2,
                                     split_train = ucan_p2_split$data_train, 
                                     split_test = ucan_p2_split$data_test))

saveRDS(ucan_p2_ml, savepath_data("ML", "ucan_p2_ml.rds"))


ovc_p2 <- do_lasso_binary(disease = "Ovarian cancer",
                          metadata = ucan_meta_p2,
                          split_train = ucan_p2_split$data_train, 
                          split_test = ucan_p2_split$data_test) 

ovc_p1$auc
ovc_p2$auc

proteins_ovc_p1 <- 
  ovc_p1$important_proteins |> 
  filter(Importance > 0)

ovc_p2$important_proteins |> filter(Importance > 0) |> 
  filter(Variable %in% proteins_ovc_p1$Variable)

```

### Combine ROC

```{r}
ucan_p1_ml[[1]]$auc
ucan_p2_ml[[1]]$auc

ucan_p1_ml[[2]]$auc
ucan_p2_ml[[2]]$auc

ucan_p1_ml[[3]]$auc
ucan_p2_ml[[3]]$auc

ucan_p1_ml[[4]]$auc
ucan_p2_ml[[4]]$auc

ucan_p1_ml[[5]]$auc
ucan_p2_ml[[5]]$auc

```
