==> scripts/06_disease_profiling_ML.Rmd <==
---
title: "04_disease_profiling_ML"
output: html_document
autor: "María Bueno Álvez"
date: "2024-12-13"
editor_options: 
  chunk_output_type: console
---

# Set up

```{r}
source("scripts/functions/functions_analyses.R")
source("scripts/functions/functions_utility.R")
source("scripts/functions/functions_visualization.R")
source("scripts/functions/themes_palettes.R")

# Read in data
data_disease <-
  read_tsv(
    "../Human-disease-blood-atlas/data/final_data/HPA/v24_2/disease_data_phase1.tsv"
  )
meta_disease <-
  read_tsv(
    "../Human-disease-blood-atlas/data/final_data/HPA/v24_2/disease_meta_phase1.tsv"
  )
```

## Disease order

```{r}
disease_class_order <-
  meta_disease |>
  distinct(Class, Disease) |>
  mutate(Class = factor(Class, levels = class_order)) |>
  arrange(Class, Disease) |>
  pull(Disease)

# Exclude diseases with more than 50 samples for machine learning
exclude_diseases <-
  meta_disease |>
  filter(DAid %in% data_disease$DAid) |>
  count(Disease) |>
  arrange(n) |>
  filter(n < 48) |>
  pull(Disease)

include_diseases <-
  meta_disease |>
  distinct(Disease) |>
  filter(!Disease %in% c(exclude_diseases, "Healthy")) |>
  pull(Disease)

disease_class_order_ml <-
  meta_disease |>
  distinct(Class, Disease) |>
  filter(Disease %in% include_diseases) |>
  mutate(Class = factor(Class, levels = class_order)) |>
  arrange(Class, Disease) |>
  pull(Disease)
```

# Prepare and import data

Identify key plasma proteins associated with disease by assessing their predictive value using machine learning. 
Generate data splits and run 100 iterations on the server.

## Generate data splits

```{r}
ml_data <-
  data_disease |>
  left_join(meta_disease |>
              select(DAid, Disease, Class), by = "DAid") |>
  filter(Disease %in% c(include_diseases, "Healthy")) |>
  select(DAid, Disease, Assay, NPX) |>
  pivot_wider(names_from = Assay,
              values_from = NPX)


# Split data 100 times
seeds <- c(1:100)
random_splits <-
  map(
    seeds,
    \(seed) generate_split(
      data = ml_data,
      proportion = 0.7,
      variable_stratify = "Disease",
      seed = seed
    )
  )

#saveRDS(random_splits, "server-scripts/data/random_splits.rds")
```

## Import parsed results from cluster 

```{r}
# Importances
importances_all <-
  read_tsv("server-scripts/parsed-results/importances_disease_all.tsv")
importances_class <-
  read_tsv("server-scripts/parsed-results/importances_disease_class.tsv")
importances_healthy <-
  read_tsv("server-scripts/parsed-results/importances_disease_healthy.tsv")

# Performance
predictions_all <-
  read_tsv("server-scripts/parsed-results/predictions_disease_all.tsv")
predictions_class <-
  read_tsv("server-scripts/parsed-results/roc_disease_class.tsv")
predictions_healthy <-
  read_tsv("server-scripts/parsed-results/roc_disease_healthy.tsv")

# ROC
roc_all <-
  read_tsv("server-scripts/parsed-results/roc_disease_all.tsv")
roc_class <-
  read_tsv("server-scripts/parsed-results/roc_disease_class.tsv")
roc_healthy <-
  read_tsv("server-scripts/parsed-results/roc_disease_healthy.tsv")
```

## Rescale & combine feature importance

```{r}
scaled_healthy <-
  importances_healthy |>
  select(
    Disease = Class,
    Control = Type,
    Seed,
    Assay = Variable,
    Importance
  ) |>
  group_by(Disease, Seed) |>
  mutate(Importance = scales::rescale(Importance, to = c(0, 100))) |>
  filter(Importance != 0)

class_scaling <-
  importances_class |>
  distinct(Disease = class, Seed) |>
  mutate(Importance = 0,
         Assay = "SCALING")

scaled_class <-
  importances_class |>
  select(
    Disease = class,
    Control = Type,
    Seed,
    Assay = term,
    Importance = estimate
  ) |>
  mutate(Importance = abs(Importance)) |>
  bind_rows(class_scaling) |>
  group_by(Disease, Seed) |>
  mutate(Importance = scales::rescale(Importance, to = c(0, 100))) |>
  filter(Assay != "SCALING")

all_scaling <-
  importances_all |>
  distinct(Disease = class, Seed) |>
  mutate(Importance = 0,
         Assay = "SCALING")

scaled_all <-
  importances_all |>
  select(
    Disease = class,
    Control = Type,
    Seed,
    Assay = term,
    Importance = estimate
  ) |>
  mutate(Importance = abs(Importance)) |>
  bind_rows(class_scaling) |>
  group_by(Disease, Seed) |>
  mutate(Importance = scales::rescale(Importance, to = c(0, 100))) |>
  filter(Assay != "SCALING")
```

## Save for HPA

```{r}
final_importance_data  <-
  scaled_healthy |>
  bind_rows(scaled_class) |>
  bind_rows(scaled_all) |>
  left_join(data_disease |>
              distinct(Assay, OlinkID), by = "Assay") |>
  relocate(OlinkID, .after = Assay) |>
  mutate(
    Disease = factor(Disease, levels = disease_class_order),
    Control = factor(Control, levels = c("All other diseases", "Class", "Healthy"))
  ) |>
  arrange(Control, Disease, Seed,-Importance)

#write_csv(final_importance_data, savepath_data("Supplementary-files", "final_importances.csv"))
#write_csv(final_importance_data, "../Human-disease-blood-atlas/data/final_data/HPA/v24_2/disease_ml.csv")
```

# Performance

Explore performance across models.

### Plot AUC comparison

```{r}
combined_auc <-
  roc_class |>
  distinct(Disease = Group, Seed, AUC) |>
  mutate(Control = "Class") |>
  bind_rows(roc_all |>
              distinct(Disease = Group, Seed, AUC) |>
              mutate(Control = "All other diseases")) |>
  bind_rows(roc_healthy |>
              distinct(Disease = Class, Seed, AUC) |>
              mutate(Control = "Healthy")) |>
  mutate(Disease = factor(Disease, levels = disease_class_order)) |>
  arrange(Disease, Seed)


all_summary <-
  combined_auc |>
  group_by(Control, Disease) |>
  summarise(mean_auc = mean(AUC),
            sd_auc = sd(AUC)) |>
  left_join(meta_disease |>
              distinct(Disease, Class)) |>
  mutate(
    Disease = factor(Disease, levels = disease_class_order),
    Class = factor(Class, levels = class_order)
  ) |>
  arrange(Disease)

all_summary |>
  ggplot(aes(Disease, mean_auc, color = Control)) +
  geom_point() +
  geom_segment(aes(
    x = Disease,
    xend = Disease,
    y = mean_auc - sd_auc,
    yend = mean_auc + sd_auc
  )) +
  scale_color_manual(values = pal_controls) +
  facet_grid( ~ Class, scales = "free_x", space = "free") +
  theme_hpa(angled = T)

ggsave(
  savepath_results("Manuscript-figures", "auc_comparison.pdf"),
  width = 12,
  height = 5
)
```
## Plot confusion matrix

```{r}
predictions_all |>
  mutate(Disease = factor(Disease),
         .pred_class = factor(.pred_class)) |>
  conf_mat(truth = Disease, estimate = .pred_class)  |>
  plot_cm(percentage = T)

ggsave(
  savepath_results("Manuscript-figures", "confusion_matrix_multiclass.pdf"),
  width = 10,
  height = 10
)
```

## Save performance

```{r}
final_performance_data <-
  combined_auc |>
  mutate(
    Disease = factor(Disease, levels = disease_class_order),
    Control = factor(Control, levels = c("All other diseases", "Class", "Healthy"))
  ) |>
  arrange(Control, Disease, Seed)

write_csv(
  final_performance_data,
  savepath_data("Supplementary-files", "final_performance.csv")
)
```
