==> scripts/08_validation_cancer_ukb.Rmd <==
---
title: "08_validation_ukb"
output: html_document
autor: "María Bueno Álvez"
date: "2024-12-13"
editor_options: 
  chunk_output_type: console
---

# Set up

```{r}
source("scripts/functions/functions_analyses.R")
source("scripts/functions/functions_utility.R")
source("scripts/functions/functions_visualization.R")
source("scripts/functions/themes_palettes.R")

# Read in data - UKB
ukb_data <- read_delim("data/UKBB_cancer_olink.txt")
data_disease <-
  read_tsv(
    "../Human-disease-blood-atlas/data/final_data/HPA/v24_2/disease_data_phase1.tsv"
  )

ukb_meta <- read_delim("data/UKBB_cancer_meta.txt")

# Extract information from data and metadata
ukb_meta <-
  ukb_meta |>
  distinct(eid, .keep_all = TRUE)

ukb_data <-
  ukb_data |>
  separate(meaning,  c("Assay", "Meaning")) |>
  select(eid, Assay, NPX = result)
```


# Time to diagnosis

Define time to diagnosis groups and visualize for each cancer type.

## Define diagnosis groups

```{r}
# Define UKB cancers
ukb_cancers <- c("Colorectal_cancer",
                 "Lung_cancer",
                 "Ovary_cancer",
                 "Prostate_cancer")

# Parse UKB metadata - define time to diagnosis groups
ukb_dat <-
  ukb_meta |>
  mutate(
    time_to_diagnosis = Age_sample_collect - Cancer_age_diagnosis,
    Cancer_name = ifelse(
      Cancer_name %in% c("Colon_cancer", "Rectum_cancer"),
      "Colorectal_cancer",
      Cancer_name
    ),
    Group = case_when(
      time_to_diagnosis <= -7 ~ "> 7 years before",
      time_to_diagnosis > -7 &
        time_to_diagnosis <= -5 ~ "5-7 years before",
      time_to_diagnosis > -5 &
        time_to_diagnosis <= -3 ~ "3-5 years before",
      time_to_diagnosis > -3 &
        time_to_diagnosis <= -1 ~ "1-3 years before",
      time_to_diagnosis > -1 &
        time_to_diagnosis <= 1 ~ "1 year before/after",
      time_to_diagnosis > 1 &
        time_to_diagnosis <= 3 ~ "1-3 years after",
      time_to_diagnosis > 3 ~ "> 3 years after"
    ),
    Group_2 = case_when(
      time_to_diagnosis <= -5 ~ "> 5 years before",
      time_to_diagnosis > -5 &
        time_to_diagnosis <= -3 ~ "3-5 years before",
      time_to_diagnosis > -3 &
        time_to_diagnosis <= -1 ~ "1-3 years before",
      time_to_diagnosis > -1 &
        time_to_diagnosis <= 1 ~ "1 year before/after",
      time_to_diagnosis > 1  ~ "> 1 year after"
    ),
    Group_3 = case_when(
      time_to_diagnosis <= -3 ~ "> 3 years before",
      time_to_diagnosis > -3 &
        time_to_diagnosis <= 0 ~ "Up to 3 years before",
      time_to_diagnosis > 0 ~ "After diagnosis"
    ),
    Group_2 = factor(Group_2, levels = names(pal_ukb_2)),
    Group_3 = factor(Group_2, levels = names(pal_ukb_3)),
    Group = factor(Group, levels = names(pal_ukb[-1]))
  ) |>
  filter(!time_to_diagnosis > 20)
```

## Visualize

```{r}
# Plot time to diagnosis distribution per cancer
ukb_dat |>
  filter(Cancer_name %in% ukb_cancers) |>
  mutate(Cancer_name = factor(
    Cancer_name,
    levels = c(
      "Lung_cancer",
      "Colorectal_cancer",
      "Prostate_cancer",
      "Ovary_cancer"
    )
  )) |>
  ggplot(aes(time_to_diagnosis, fill = Group_2)) +
  geom_histogram() +
  geom_vline(xintercept = 0) +
  scale_fill_manual(values = pal_ukb_2) +
  facet_wrap(~ Cancer_name, nrow = 1, scales = "free_y") +
  theme_hpa() +
  ggtitle("") +
  xlab("Time to diagnosis")

ggsave(savepath_results("Manuscript-figures", "age_ukb.pdf"),
       h = 3.5,
       w = 12)
```
# Machine learning 

Test models developed in the U-CAN cohort for lung, colorectal, ovarian and prostate cancer in several UKB diagnosis groups.

## Load U-CAN models

```{r}
# Define U - CAN cancers
cancers <- c("Lung cancer",
             "Colorectal cancer",
             "Prostate cancer",
             "Ovarian cancer")

# Load U-CAN model results
models_ucan_p1 <- readRDS(savepath_data("ML", "models_ucan_p1.rds"))
```

## Prepare UKB data

```{r}
ukb_wide <-
  ukb_data |>
  left_join(ukb_dat, by = "eid") |>
  select(DAid = eid, Assay, NPX, Cancer_name) |>
  mutate(
    Disease = case_when(
      Cancer_name == "Colorectal_cancer" ~ "Colorectal cancer",
      Cancer_name == "Lung_cancer" ~ "Lung cancer",
      Cancer_name == "Prostate_cancer" ~ "Prostate cancer",
      Cancer_name == "Ovary_cancer" ~ "Ovarian cancer"
    )
  ) |>
  select(-Cancer_name) |>
  filter(Assay != "HLA") |> # Formatting issue
  pivot_wider(names_from = Assay,
              values_from = NPX,
              values_fn = first) |>
  filter(!is.na(Disease)) |>
  mutate(DAid = as.character(DAid))

# Identify non-mapping proteins Olink 1.5K to 3K
missing_3072 <-
  data_disease |>
  distinct(Assay) |>
  filter(!Assay %in% unique(ukb_data$Assay)) |>
  pull(Assay)

ukb_wide[missing_3072] <- NA
```

## Run predictions

```{r}
predictions_ukb <-
  map_df(cancers, function(cancer) {
    
    current_data <-
      ukb_wide |>
      mutate(Variable = case_when(
        Disease == cancer ~ paste0("1_", cancer),
        Disease != cancer ~ paste0("0_Control")
      )) |>
      select(-Disease) |>
      mutate(Variable = factor(Variable))
    
    if (cancer %in% female_diseases) {
      current_data <-
        current_data |>
        left_join(ukb_dat |>
                    mutate(eid = as.character(eid)) |>
                    select(eid, Sex),
                  by = c("DAid" = "eid")) |>
        filter(Sex == 0) |>
        select(-Sex)
    } else if (cancer %in% male_diseases) {
      current_data <-
        current_data |>
        left_join(ukb_dat |>
                    mutate(eid = as.character(eid)) |>
                    select(eid, Sex),
                  by = c("DAid" = "eid")) |>
        filter(Sex == 1) |>
        select(-Sex)
    } else {
      current_data <- current_data
    }
    
    current_model <- models_ucan_p1[[cancer]]$model
    
    new_predictions <-
      predict(current_model, new_data = current_data, type = "prob")
    
    group_info <-
      ukb_wide |>
      filter(Disease == cancer) |>
      left_join(ukb_dat |>
                  select(DAid = eid, Group_2) |>
                  mutate(DAid = as.character(DAid)),
                by = "DAid") |>
      select(DAid, Group_2)
    
    new_predictions |>
      mutate(DAid = current_data$DAid) |>
      left_join(group_info) |>
      mutate(Variable = current_data$Variable,
             Cancer = cancer)
    
  })

saveRDS(predictions_ukb, savepath_data("ML", "predictions_ukb.rds"))
```
 
## Change granularity (diagnosis groups)

```{r}
all_pred <-
  map_df(cancers, function(cancer) {
    new_predictions <-
      predictions_ukb |>
      filter(Cancer == cancer)
    

    map_df(as.character(unique(group_info$Group_2)), function(group) {
      new_predictions_ext <-
        new_predictions |>
        left_join(group_info)
      
      current_roc <-
        new_predictions_ext |>
        filter(Variable == "0_Control" | Group_2 == group) |>
        mutate(Variable = factor(Variable)) |>
        roc_curve(truth = Variable,
                  paste0(".pred_1_", cancer),
                  event_level = "second") |>
        mutate(Group = group)
      
      new_ext_res <-
        new_predictions_ext |>
        filter(Variable == "0_Control" | Group_2 == group) |>
        mutate(Variable = factor(Variable)) |>
        rename(prediction = paste0(".pred_1_", cancer))
      
      auc <- pROC::auc(new_ext_res$Variable, new_ext_res$prediction)
      ci <-
        pROC::ci.auc(new_ext_res$Variable, new_ext_res$prediction)
      
      # Combine
      combined_roc_auc <-
        current_roc |>
        mutate(
          AUC = as.numeric(auc),
          CI_lower = as.numeric(ci)[[1]],
          CI_upper = as.numeric(ci)[[3]],
          Cancer = cancer
        )
    })
  })        
```

## ROC

```{r}
tiles <-
  all_pred |>
  mutate(Cancer = factor(Cancer, levels = cancers),
         Group = factor(Group, levels = names(pal_ukb_2))) |>
  distinct(Group, Cancer, AUC) |>
  mutate(
    specificity = case_when(
      Group == "> 5 years before" ~ 0.1,
      Group == "3-5 years before" ~ 0.1,
      Group == "1-3 years before" ~ 0.1,
      Group == "1 year before/after" ~ 0.1,
      Group == "> 1 year after" ~ 0.1
    ),
    sensitivity = case_when(
      Group == "> 5 years before" ~ 0.15,
      Group == "3-5 years before" ~ 0.2,
      Group == "1-3 years before" ~ 0.25,
      Group == "1 year before/after" ~ 0.3,
      Group == "> 1 year after" ~ 0.35
    ),
    Cancer = factor(Cancer, levels = rev(cancers)),
    Group = factor(Group, levels = names(pal_ukb_2))
  )


all_pred |>
  mutate(Cancer = factor(Cancer, levels = cancers),
         Group = factor(Group, levels = names(pal_ukb_2))) |>
  ggplot(aes(
    x = 1 - specificity,
    y = sensitivity,
    color = Group
  )) +
  geom_line(size = 0.8, alpha = 0.6) +  # Lighter ROC curves
  annotate(
    "segment",
    x = 0,
    y = 0,
    xend = 1,
    yend = 1,
    colour = 'grey50',
    linetype = 'dashed',
    size = 0.8
  )  +
  geom_tile(
    data = tiles,
    aes(fill = Group),
    width = 0.1,
    height = 0.05,
    show.legend = FALSE
  ) +
  geom_text(
    data = tiles,
    aes(label = round(AUC, 2)),
    size = 3,
    color = "white",
    fontface = "bold",
    show.legend = FALSE
  ) +
  facet_wrap( ~ Cancer, nrow = 1, scales = "free_x") +
  scale_x_continuous(breaks = c(0, 1)) +
  scale_y_continuous(breaks = c(0, 1)) +
  scale_color_manual(values = pal_ukb_2) +
  scale_fill_manual(values = pal_ukb_2) +
  theme_hpa() +
  theme(axis.text = element_text(size = 8),
        legend.position = "top")

ggsave(savepath_results("Manuscript-figures", "UKB_ML.pdf"),
       h = 4,
       w = 9) 
```


## Save performance

```{r}
all_pred |>
  distinct(Cancer, Group, AUC, CI_lower, CI_upper) |>
  write_csv(savepath_data("Supplementary-files", "UKB_ML_performance.csv"))
```

# Plot two examples

Generate boxplots for two example proteins across diagnosis groups.

```{r}
p_ceacam <-
  plot_boxplot_ukb("Lung_cancer", "CEACAM5")  +
  theme_hpa(axis_x = F)

p_wfdc2 <-
  plot_boxplot_ukb("Ovary_cancer", "WFDC2") +
  xlab("Time to diagnosis groups")

p_ceacam / p_wfdc2

ggsave(
  savepath_results("Manuscript-figures", "UKB_examples.pdf"),
  h = 7,
  w = 4
)
```
